refactoring TF code and moving it out of App.tsx into TF service
